Transition cvp.opnfv.org to verified.opnfv.org
As cvp is being renamed to the OPNFV Verified program, the url of the
dashboard site needs to be updated, but before we fully switch over both
domains should still resolve.

The DNS record for verified.opnfv.org is in place and this change has
been verified to work on the server.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>